LPD-87401 Move email test into SharingEntryLocalServiceTest
