--- modules/apps/sharing/sharing-test/src/testIntegration/java/com/liferay/sharing/notifications/test/SharingEntryCollaborationEmailTest.java
@@ -1,173 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.sharing.notifications.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.portal.kernel.model.Company;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.Ticket;
-import com.liferay.portal.kernel.model.TicketConstants;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
-import com.liferay.portal.kernel.service.ClassNameLocalService;
-import com.liferay.portal.kernel.service.CompanyLocalService;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.TicketLocalService;
-import com.liferay.portal.kernel.test.TestInfo;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.test.util.UserTestUtil;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.PrefsPropsUtil;
-import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.test.mail.MailMessage;
-import com.liferay.portal.test.mail.MailServiceTestUtil;
-import com.liferay.portal.test.rule.FeatureFlag;
-import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.SynchronousMailTestRule;
-import com.liferay.sharing.security.permission.SharingEntryAction;
-import com.liferay.sharing.service.SharingEntryLocalService;
-
-import java.util.Arrays;
-import java.util.Date;
-import java.util.concurrent.TimeUnit;
-
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import org.springframework.mock.web.MockHttpServletRequest;
-
-/**
- * @author Alicia García
- */
-@FeatureFlag("LPD-52006")
-@RunWith(Arquillian.class)
-public class SharingEntryCollaborationEmailTest {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(), SynchronousMailTestRule.INSTANCE);
-
-	@Before
-	public void setUp() throws Exception {
-		_group = GroupTestUtil.addGroup();
-
-		_company = _companyLocalService.getCompany(_group.getCompanyId());
-
-		_fromUser = UserTestUtil.addUser();
-	}
-
-	@Test
-	@TestInfo("LPD-48130")
-	public void testAddSharingEntryWithInviteCollaboratorTicket()
-		throws Exception {
-
-		String emailAddress = RandomTestUtil.randomString() + "@liferay.com";
-
-		Ticket ticket = _ticketLocalService.addTicket(
-			TestPropsValues.getCompanyId(), Group.class.getName(),
-			_group.getGroupId(), TicketConstants.TYPE_INVITE_COLLABORATOR,
-			emailAddress,
-			new Date(System.currentTimeMillis() + TimeUnit.HOURS.toMillis(48)),
-			new ServiceContext());
-
-		MailServiceTestUtil.clearMessages();
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_group.getGroupId(), _fromUser.getUserId());
-
-		serviceContext.setRequest(_createMockHttpServletRequest());
-
-		_sharingEntryLocalService.addSharingEntry(
-			null, _fromUser.getUserId(), ticket.getTicketId(), 0, 0,
-			_classNameLocalService.getClassNameId(Group.class.getName()),
-			_group.getGroupId(), _group.getGroupId(), true,
-			Arrays.asList(SharingEntryAction.VIEW), null, serviceContext);
-
-		Assert.assertEquals(1, MailServiceTestUtil.getInboxSize());
-
-		MailMessage mailMessage = MailServiceTestUtil.getLastMailMessage();
-
-		String body = mailMessage.getBody();
-
-		String fromAddress = PrefsPropsUtil.getString(
-			_company.getCompanyId(), PropsKeys.ADMIN_EMAIL_FROM_ADDRESS);
-		String fromName = PrefsPropsUtil.getString(
-			_company.getCompanyId(), PropsKeys.ADMIN_EMAIL_FROM_NAME);
-
-		Assert.assertTrue(body.contains("create_account_user"));
-		Assert.assertTrue(body.contains(fromAddress));
-		Assert.assertTrue(body.contains(fromName));
-		Assert.assertTrue(body.contains(ticket.getKey()));
-
-		String from = mailMessage.getFirstHeaderValue("From");
-
-		Assert.assertTrue(from.contains(fromAddress));
-
-		String subject = mailMessage.getFirstHeaderValue("Subject");
-
-		Assert.assertTrue(subject.contains(fromName));
-
-		String to = mailMessage.getFirstHeaderValue("To");
-
-		Assert.assertTrue(to.contains(emailAddress));
-	}
-
-	private MockHttpServletRequest _createMockHttpServletRequest()
-		throws Exception {
-
-		MockHttpServletRequest mockHttpServletRequest =
-			new MockHttpServletRequest();
-
-		ThemeDisplay themeDisplay = new ThemeDisplay();
-
-		themeDisplay.setCompany(_company);
-		themeDisplay.setPermissionChecker(
-			PermissionCheckerFactoryUtil.create(_fromUser));
-		themeDisplay.setScopeGroupId(_group.getGroupId());
-		themeDisplay.setSiteGroupId(_group.getGroupId());
-		themeDisplay.setUser(_fromUser);
-
-		mockHttpServletRequest.setAttribute(
-			WebKeys.THEME_DISPLAY, themeDisplay);
-
-		return mockHttpServletRequest;
-	}
-
-	@Inject
-	private ClassNameLocalService _classNameLocalService;
-
-	private Company _company;
-
-	@Inject
-	private CompanyLocalService _companyLocalService;
-
-	@DeleteAfterTestRun
-	private User _fromUser;
-
-	@DeleteAfterTestRun
-	private Group _group;
-
-	@Inject
-	private SharingEntryLocalService _sharingEntryLocalService;
-
-	@Inject
-	private TicketLocalService _ticketLocalService;
-
-}
